+++ src/stdlib/crypto_async.rs
@@ -40,7 +40,10 @@ use zeroize::Zeroizing;
 ///
 /// `constant_time_equals` and `secure_random_bytes` are intentionally *not*
 /// routed: they are fast and routing them would only add a scheduling hop.
-pub(crate) fn route(
+///
+/// Public so the interpreter dispatch (and the `tests/crypto_async_test.rs`
+/// integration tests) can drive it; not part of the stable language surface.
+pub fn route(
     name: &str,
     args: &[Value],
 ) -> Option<LocalBoxFuture<'static, Result<Value, RuntimeError>>> {
@@ -149,159 +152,4 @@ fn extract_password_and_stored(
 /// join error must not be silently swallowed.
 fn join_error(func: &str, e: tokio::task::JoinError) -> RuntimeError {
     RuntimeError::new(format!("{func}: crypto task failed: {e}"), 0, 0)
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use std::sync::atomic::{AtomicUsize, Ordering};
-
-    fn text(s: &str) -> Value {
-        Value::Text(Arc::from(s))
-    }
-
-    /// Only the deliberately-slow crypto builtins route; everything else falls
-    /// through to the synchronous native. Guards against silent scope creep in
-    /// the routed set.
-    #[test]
-    fn route_map_covers_exactly_the_heavy_crypto_builtins() {
-        let one = [text("x")];
-        let two = [text("x"), text("y")];
-        let four = [
-            text("x"),
-            text("y"),
-            Value::Number(1.0),
-            Value::Number(16.0),
-        ];
-
-        for name in [
-            "argon2_hash",
-            "scrypt_hash",
-            "bcrypt_hash",
-            "pbkdf2_hash",
-            "hash_password",
-        ] {
-            assert!(route(name, &one).is_some(), "{name} should route");
-        }
-        for name in [
-            "argon2_verify",
-            "scrypt_verify",
-            "bcrypt_verify",
-            "pbkdf2_verify",
-            "verify_password",
-        ] {
-            assert!(route(name, &two).is_some(), "{name} should route");
-        }
-        assert!(route("pbkdf2_hmac_sha256", &four).is_some());
-
-        // Fast / non-crypto builtins must NOT route.
-        for name in [
-            "constant_time_equals",
-            "secure_random_bytes",
-            "sha256",
-            "wflhash256",
-            "print",
-            "length",
-        ] {
-            assert!(route(name, &two).is_none(), "{name} must not route");
-        }
-    }
-
-    /// The core Phase 0 property (plan U4): the heavy work runs off the
-    /// interpreter thread. On a single-threaded runtime, a concurrently-spawned
-    /// ticker can only make progress *during* the crypto `.await` if that work
-    /// was offloaded via `spawn_blocking`. Had the hash run inline, it would
-    /// have monopolized the one thread and the ticker would still read zero when
-    /// the await returns. This is deterministic and independent of core count.
-    #[tokio::test(flavor = "current_thread")]
-    async fn routed_crypto_frees_the_interpreter_thread() {
-        let counter = std::sync::Arc::new(AtomicUsize::new(0));
-        let ticker_counter = counter.clone();
-        let ticker = tokio::spawn(async move {
-            for _ in 0..5 {
-                tokio::task::yield_now().await;
-                ticker_counter.fetch_add(1, Ordering::SeqCst);
-            }
-        });
-
-        let args = [text("a-password-to-hash")];
-        let out = route("argon2_hash", &args)
-            .expect("argon2_hash routes")
-            .await
-            .expect("hash succeeds");
-
-        // Read the ticker's progress the instant the await returns, before we
-        // yield again — so a passing result can only mean the runtime polled the
-        // ticker while the blocking pool computed the hash.
-        let progressed = counter.load(Ordering::SeqCst);
-        ticker.await.expect("ticker joins");
-
-        assert!(
-            progressed > 0,
-            "ticker never advanced during the crypto await — work was not offloaded"
-        );
-        assert!(matches!(out, Value::Text(_)));
-    }
-
-    /// The routed hash still produces a valid, verifiable credential.
-    #[tokio::test]
-    async fn routed_argon2_hash_roundtrips() {
-        let out = route("argon2_hash", &[text("correct horse")])
-            .unwrap()
-            .await
-            .unwrap();
-        let hash = match out {
-            Value::Text(h) => h,
-            other => panic!("expected text, got {other:?}"),
-        };
-        assert!(crypto::argon2_verify_str("correct horse", &hash));
-        assert!(!crypto::argon2_verify_str("wrong password", &hash));
-    }
-
-    /// bcrypt uses a distinct MCF (`$2b$`) format, exercising a different helper.
-    #[tokio::test]
-    async fn routed_bcrypt_hash_roundtrips() {
-        let out = route("bcrypt_hash", &[text("hunter2")])
-            .unwrap()
-            .await
-            .unwrap();
-        let hash = match out {
-            Value::Text(h) => h,
-            other => panic!("expected text, got {other:?}"),
-        };
-        assert!(hash.starts_with("$2"));
-        assert!(crypto::bcrypt_verify_str("hunter2", &hash));
-        assert!(!crypto::bcrypt_verify_str("nope", &hash));
-    }
-
-    /// The routed KDF output is byte-identical to the direct plain-data helper —
-    /// the thread hop changes nothing about the result (deterministic: no salt).
-    #[tokio::test]
-    async fn routed_pbkdf2_matches_direct() {
-        let args = [
-            text("password"),
-            text("salt"),
-            Value::Number(4096.0),
-            Value::Number(32.0),
-        ];
-        let out = route("pbkdf2_hmac_sha256", &args).unwrap().await.unwrap();
-        let routed = match out {
-            Value::Text(h) => h.to_string(),
-            other => panic!("expected text, got {other:?}"),
-        };
-        let direct = crypto::pbkdf2_hmac_sha256_str("password", "salt", 4096, 32).unwrap();
-        assert_eq!(routed, direct);
-    }
-
-    /// Argument errors still surface (as the routed future's error), so routing
-    /// does not mask misuse.
-    #[tokio::test]
-    async fn routed_argerror_is_reported() {
-        // argon2_hash expects exactly one argument.
-        let err = route("argon2_hash", &[text("a"), text("b")])
-            .unwrap()
-            .await
-            .unwrap_err();
-        assert!(err.message.contains("argon2_hash"));
-    }
 }
